RichText: change withFormatTypes to hook
